chore(refactor): refactored the home page
