+++ src/components/Home/index.js
@@ -1,0 +1,4 @@
+export { default as HeroSection } from './HeroSection';
+export { default as InfoSection } from './InfoSection';
+export { default as FeaturesSection } from './FeaturesSection';
+export { default as HowItWorksSection } from './HowItWorksSection';
